+++ CHANGELOG.md
@@ -21,6 +21,7 @@ changes. Best viewed [here](https://google-grain.readthedocs.io/en/latest/change
   * Introduces `grain.experimental.batch_and_pad` for padding a partial batch to
     avoid dropping batch remainder data.
   * Upgrades `array_record` and `protobuf`.
+  * Make keyfile optional when only doing lookup-by-key with sstable sources
 
 * Breaking changes: