+++ CHANGELOG.md
@@ -5,14 +5,24 @@ changes. Best viewed [here](https://google-grain.readthedocs.io/en/latest/change
 
 ## Unreleased
 
+* New features
+
+* Breaking changes:
+
+* Deprecations:
+
+* Bug fixes
+
+## Grain 0.2.12 (August 21, 2025)
+
 * New features:
+  * Adds Windows build.
   * Allow passing `read_kwargs` to `ParquetIterDataset` for configuring parquet
     file reading.
   * `ThreadPrefetchDatasetIterator` now supports non-Grain iterators that
     support checkpointing.
   * Introduces API for device prefetch - `grain.experimental.device_put()` for
     easy CPU and device prefetching.
-  * Adds Windows build.
   * Introduces API for autotuning -- given the user provided RAM restrictions
     and specific `IterDataset`, finds number of processes for `mp_prefetch`
     and buffer size for `PrefetchDatasetIterator`.
@@ -20,11 +30,11 @@ changes. Best viewed [here](https://google-grain.readthedocs.io/en/latest/change
     array record file reading.
   * Introduces `grain.experimental.batch_and_pad` for padding a partial batch to
     avoid dropping batch remainder data.
-  * Upgrades `array_record` and `protobuf`.
   * Grain interleave optimization - allow creating more threads to parallelly
     keep starting iterators and prefetching elements.
 
 * Breaking changes:
+  * Upgrades `array_record` and `protobuf`.
 
 * Deprecations: